legal_spend/requirements.txt
google-adk
pandas
openpyxl
ollama
